(PC-20410) refacto(adaptGeolocationParameters): extract adaptGeolocationParameters into a fetchAlgolia helper
